move chrome trace flushing for blocking OpenCL APIs
For blocking OpenCL APIs, flush any accumulated chrome trace records
before issuing the blocking OpenCL API.  In cases where the blocking
OpenCL API takes a long time to execute, this can reduce the effective
overhead of flushing the chrome trace buffer to zero.

This isn't perfect, and there are some downsides to this approach:

The largest downside occurs when there are a large number of chrome
trace records to flush, or when the blocking API finishes promptly.
When this occurs, the reported time for the blocking API may be larger
than it would otherwise, because it includes the chrome trace flushing
overhead.

Additionally, because the chrome trace flushing overhead is also
tracked as part of the time for the blocking API call, it is essentially
double counted.  In cases when it is important to track the total host
time, the double counting can be eliminated by disabling tool overhead
tracking.